+++ src/MEDWrapper/V2_1/Core/MEDchampEcr.cxx
@@ -177,8 +177,8 @@ MEDchampEcr(med_idt fid, char *maa, char *cha,unsigned char *val,med_mode_switch
     else
       psize = i;
     
-    pfltabtmp = (med_int *)   malloc (sizeof(med_int)*psize);
-    pfltab    = (med_ssize *) malloc (sizeof(med_ssize)*psize);
+    pfltabtmp = (med_int *)   malloc (sizeof(med_int)*(size_t)psize);
+    pfltab    = (med_ssize *) malloc (sizeof(med_ssize)*(size_t)psize);
     if ((ret = MEDprofilLire(fid,pfltabtmp,pflname)) < 0)
       return -1;
     for (i=0;i<psize;i++)
